requirements.txt
streamlit>=1.28.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
bcrypt>=4.0.0
pandas>=2.0.0
plotly>=5.0.0
python-dotenv>=1.0.0
reportlab>=4.0.0
openpyxl>=3.1.0
Pillow>=10.0.0
supabase>=2.0.0
google-auth>=2.0.0
requests>=2.28.0
